requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Werkzeug==2.3.7
SQLAlchemy==2.0.21
requests==2.31.0
pytube==12.1.0
python-dotenv==1.0.0
Jinja2==3.0.1
mysqlclient==2.2.0
PyMySQL==1.0.2 